@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Http, Headers, RequestOptions, Response } from '@angular/http';
+import { HttpClient, HttpHeaders, HttpRequest, HttpProgressEvent } from '@angular/common/http'
 import { Observable } from 'rxjs/Observable';
 import { Movie } from './dvd';
 import { Angular2TokenService } from 'angular2-token';
@@ -7,10 +8,14 @@ import 'rxjs/add/operator/map';
 
 @Injectable()
 export class DvdService {
+	// moviesURL = 'http://localhost:3001/movies';
 	moviesURL = 'http://moviedatabase-env.us-west-2.elasticbeanstalk.com/movies';
 
-	constructor(private http: Http,
-							private authTokenService: Angular2TokenService) { }
+	constructor(
+		private http: Http,
+		private httpClient: HttpClient,
+		private authTokenService: Angular2TokenService
+	) { }
 
 	getMovies(): Observable<Movie[]> {
 		const headers = new Headers();
@@ -28,6 +33,7 @@ export class DvdService {
 		headers.append('access-token', this.authTokenService.currentAuthData.accessToken);
 		headers.append('client', this.authTokenService.currentAuthData.client);
 		const options = new RequestOptions({ headers: headers });
+		const req = new XMLHttpRequest
 
 		return this.http.get(this.moviesURL + '/' + id + '.mp4', options)
 			.map(
@@ -36,17 +42,23 @@ export class DvdService {
 			)
 	}
 
-	newMovie(fileToUpload: File, form): Observable<Movie> {
+	newMovie(fileToUpload: File, title: string, year: string, plot: string) {
 		const formData = new FormData();
-		formData.append('title', form.title);
-		formData.append('year', form.year);
-		formData.append('plot', form.plot);
+		formData.append('title', title);
+		formData.append('year', year);
+		formData.append('plot', plot);
 		formData.append('video', fileToUpload);
 		const headers = new Headers();
 		headers.delete('Content-Type');
 		headers.append('access-token', this.authTokenService.currentAuthData.accessToken);
 		headers.append('client', this.authTokenService.currentAuthData.client);
 		const options = new RequestOptions({ headers: headers });
+
+		// const req = new HttpRequest<Movie>('POST', this.moviesURL, formData, {
+		// 	headers: headers,
+		// 	reportProgress: true
+		// })
+		// return this.httpClient.request(req)
 
 		return this.http.post(this.moviesURL, formData, options)
 			.map((res) => res.json())
