@@ -1,8 +1,8 @@
-import { OrderbyPipe } from './orderby.pipe';
+import { OrderByPipe } from './orderby.pipe';
 
 describe('OrderbyPipe', () => {
-  it('create an instance', () => {
-    const pipe = new OrderbyPipe();
-    expect(pipe).toBeTruthy();
-  });
+	it('create an instance', () => {
+		const pipe = new OrderByPipe();
+		expect(pipe).toBeTruthy();
+	});
 });
