@@ -5,6 +5,6 @@ export const emptyEmail = (control: AbstractControl) => {
 	if (!email.value) {
 		email.setValidators([])
 	} else {
-		email.setValidators([Validators.email, Validators.pattern("[a-zA-Z0-9.-_]{1,}@[a-zA-Z.-]{2,}[.]{1}[a-zA-Z]{2,}")]);
+		email.setValidators([Validators.email, Validators.pattern('[a-zA-Z0-9.-_]{1,}@[a-zA-Z.-]{2,}[.]{1}[a-zA-Z]{2,}')]);
 	}
 }
